v0.3.3

+++ BeingAliveLanguageCppPort/cppFunc.cpp
@@ -1,20 +1,20 @@
 #include "stdafx.h"
 #include "cppFunc.h"
 
-void BAL_possionDiskElimSample(ON_SimpleArray<float>* inPt, double area, int n,
+void BAL_possionDiskElimSample(ON_SimpleArray<double>* inPt, double area, int n,
 	ON_3dPointArray* outPt) {
 	// input conversion
 	int sz = inPt->Count() / 3;
 
-	std::vector<cy::Vec3f> inputPoints(0);
+	std::vector<cy::Vec3d> inputPoints(0);
 	for (size_t i = 0; i < sz; i++) {
 		inputPoints.emplace_back(*inPt->At(i * 3), *inPt->At(i * 3 + 1),
 			*inPt->At(i * 3 + 2));
 	}
 
 	// elimination to the given number of pts
-	cy::WeightedSampleElimination<cy::Vec3f, float, 3> wse;
-	std::vector<cy::Vec3f> outputPoints(n);
+	cy::WeightedSampleElimination<cy::Vec3d, double, 3> wse;
+	std::vector<cy::Vec3d> outputPoints(n);
 
 	//! Important!
 	// d_max is used to define the sampling dist param based on sampling area
@@ -31,6 +31,7 @@ void BAL_possionDiskElimSample(ON_SimpleArray<float>* inPt, double area, int n,
 		outPt->Append(ON_3dPoint(p.x, p.y, p.z));
 	}
 }
+
 
 double BAL_Addition(double a, double b)
 {
